requirements.txt
anyio==3.7.1
Authlib==1.2.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.25.0
cryptography==41.0.1
dnspython==2.4.0
exceptiongroup==1.2.0
Flask==2.3.2
Flask-PyMongo==2.3.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
pymongo==4.4.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.17.0
